Split out `match_ref_pats`
